+++ docs/reference/changelog.rst
@@ -45,6 +45,8 @@ Models
 - ``any`` and ``*`` are now reserved names for platforms.
 - Platform names may not contain '/' characters.
 - Ubuntu Questing is now a 'stable' base, and Resolute is added as the new 'devel' one.
+- Add a classmethod on the :py:class:`~craft_application.models.Project` class to let
+  applications fine-tune which bases are considered 'devel'.
 
 Fetch Service
 ~~~~~~~~~~~~~